+++ src/types/char.rs
@@ -163,8 +163,7 @@ fn from_code_nv(args: &[NanValue], arena: &mut Arena) -> Result<NanValue, Runtim
             let s = c.to_string();
             let s_idx = arena.push_string(&s);
             let inner = NanValue::new_string(s_idx);
-            let box_idx = arena.push_boxed(inner);
-            Ok(NanValue::new_some(box_idx))
+            Ok(NanValue::new_some_value(inner, arena))
         }
         None => Ok(NanValue::NONE),
     }